+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/ClickableIngredient.java
@@ -1,5 +1,6 @@
 package com.refinedmods.refinedstorage.jei.common;
 
+import mezz.jei.api.ingredients.IIngredientType;
 import mezz.jei.api.ingredients.ITypedIngredient;
 import mezz.jei.api.runtime.IClickableIngredient;
 import net.minecraft.client.renderer.Rect2i;
@@ -14,8 +15,19 @@ class ClickableIngredient<T> implements IClickableIngredient<T> {
     }
 
     @Override
+    @SuppressWarnings({"removal"})
     public ITypedIngredient<T> getTypedIngredient() {
         return ingredient;
+    }
+
+    @Override
+    public IIngredientType<T> getIngredientType() {
+        return ingredient.getType();
+    }
+
+    @Override
+    public T getIngredient() {
+        return ingredient.getIngredient();
     }
 
     @Override