+++ refinedstorage-jei-integration-common/src/main/java/com/refinedmods/refinedstorage/jei/common/PatternGridSmithingTableRecipeTransferHandler.java
@@ -13,14 +13,13 @@ import mezz.jei.api.gui.ingredient.IRecipeSlotsView;
 import mezz.jei.api.recipe.RecipeIngredientRole;
 import mezz.jei.api.recipe.RecipeType;
 import mezz.jei.api.recipe.transfer.IRecipeTransferError;
-import mezz.jei.api.recipe.transfer.IRecipeTransferHandler;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.MenuType;
 import net.minecraft.world.item.crafting.RecipeHolder;
 import net.minecraft.world.item.crafting.SmithingRecipe;
 
-class PatternGridSmithingTableRecipeTransferHandler implements
-    IRecipeTransferHandler<PatternGridContainerMenu, RecipeHolder<SmithingRecipe>> {
+class PatternGridSmithingTableRecipeTransferHandler
+    extends AbstractPatternGridRecipeTransferHandler<RecipeHolder<SmithingRecipe>> {
     @Override
     public Class<? extends PatternGridContainerMenu> getContainerClass() {
         return PatternGridContainerMenu.class;
@@ -55,6 +54,6 @@ class PatternGridSmithingTableRecipeTransferHandler implements
                 inputSlots.get(2)
             );
         }
-        return null;
+        return autocraftableHint(containerMenu.getView(), recipeSlots);
     }
 }